requirements.txt
miflora~=0.6
bluepy==1.3.0
btlewrap~=0.0.8
python-telegram-bot==12.7
PyYAML~=5.3.1